change useActiveElement to activeElement
