+++ packages/runed/src/lib/functions/activeElement/index.ts
@@ -1,0 +1,1 @@
+export * from "./activeElement.svelte.js";
