+++ packages/runed/src/lib/functions/index.ts
@@ -1,5 +1,5 @@
 export * from "./box/index.js";
-export * from "./useActiveElement/index.js";
+export * from "./activeElement/index.js";
 export * from "./useDebounce/index.js";
 export * from "./useElementSize/index.js";
 export * from "./useEventListener/index.js";
