--- packages/runed/src/lib/functions/useActiveElement/index.ts
@@ -1,1 +1,0 @@
-export * from "./useActiveElement.svelte.js";
